+++ datatypes/src/main/java/org/hyperledger/besu/datatypes/Blob.java
@@ -1,5 +1,5 @@
 /*
- * Copyright contributors to Hyperledger Besu.
+ * Copyright contributors to Besu.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -14,66 +14,18 @@
  */
 package org.hyperledger.besu.datatypes;
 
-import org.hyperledger.besu.ethereum.rlp.RLPInput;
-import org.hyperledger.besu.ethereum.rlp.RLPOutput;
-
-import java.util.Objects;
-
 import org.apache.tuweni.bytes.Bytes;
 
-/** Arbitrary data for use in the KZG scheme. */
-public class Blob {
-
-  final Bytes data;
-
-  /**
-   * Create a new Blob.
-   *
-   * @param data that represents the blob.
-   */
-  public Blob(final Bytes data) {
-    this.data = data;
-  }
-
-  /**
-   * Read a Blob from an RLPInput.
-   *
-   * @param input to read from.
-   * @return the Blob.
-   */
-  public static Blob readFrom(final RLPInput input) {
-    final Bytes bytes = input.readBytes();
-    return new Blob(bytes);
-  }
-
-  /**
-   * Write the Blob to an RLPOutput.
-   *
-   * @param out to write to.
-   */
-  public void writeTo(final RLPOutput out) {
-    out.writeBytes(data);
-  }
+/**
+ * Represents an arbitrary data structure used in the KZG scheme. This interface defines the
+ * contract for a Blob, which contains data.
+ */
+public interface Blob {
 
   /**
    * Get the data of the Blob.
    *
    * @return the data.
    */
-  public Bytes getData() {
-    return data;
-  }
-
-  @Override
-  public boolean equals(final Object o) {
-    if (this == o) return true;
-    if (o == null || getClass() != o.getClass()) return false;
-    Blob blob = (Blob) o;
-    return Objects.equals(getData(), blob.getData());
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(getData());
-  }
+  Bytes getData();
 }
