+++ datatypes/src/main/java/org/hyperledger/besu/datatypes/KZGCommitment.java
@@ -1,5 +1,5 @@
 /*
- * Copyright contributors to Hyperledger Besu.
+ * Copyright contributors to Besu.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -14,65 +14,17 @@
  */
 package org.hyperledger.besu.datatypes;
 
-import org.hyperledger.besu.ethereum.rlp.RLPInput;
-import org.hyperledger.besu.ethereum.rlp.RLPOutput;
+import org.apache.tuweni.bytes.Bytes;
 
-import java.util.Objects;
-
-import org.apache.tuweni.bytes.Bytes48;
-
-/** This class contains the data for a KZG commitment. */
-public class KZGCommitment {
-  final Bytes48 data;
-
+/**
+ * Represents a KZG commitment, which contains data used in the KZG scheme. This interface defines
+ * the contract for a KZG commitment, which provides access to its data.
+ */
+public interface KZGCommitment {
   /**
-   * Constructor for a KZG commitment.
+   * Get the data of the Blob.
    *
-   * @param data The data for the KZG commitment.
+   * @return the data.
    */
-  public KZGCommitment(final Bytes48 data) {
-    this.data = data;
-  }
-
-  /**
-   * Reads a KZG commitment from the RLP input.
-   *
-   * @param input The RLP input.
-   * @return The KZG commitment.
-   */
-  public static KZGCommitment readFrom(final RLPInput input) {
-    final Bytes48 bytes = input.readBytes48();
-    return new KZGCommitment(bytes);
-  }
-
-  /**
-   * Writes the KZG commitment to the RLP output.
-   *
-   * @param out The RLP output.
-   */
-  public void writeTo(final RLPOutput out) {
-    out.writeBytes(data);
-  }
-
-  /**
-   * Gets the data for the KZG commitment.
-   *
-   * @return The data for the KZG commitment.
-   */
-  public Bytes48 getData() {
-    return data;
-  }
-
-  @Override
-  public boolean equals(final Object o) {
-    if (this == o) return true;
-    if (o == null || getClass() != o.getClass()) return false;
-    KZGCommitment that = (KZGCommitment) o;
-    return Objects.equals(getData(), that.getData());
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(getData());
-  }
+  Bytes getData();
 }
