+++ datatypes/src/main/java/org/hyperledger/besu/datatypes/KZGProof.java
@@ -1,5 +1,5 @@
 /*
- * Copyright contributors to Hyperledger Besu.
+ * Copyright contributors to Besu.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -14,65 +14,17 @@
  */
 package org.hyperledger.besu.datatypes;
 
-import org.hyperledger.besu.ethereum.rlp.RLPInput;
-import org.hyperledger.besu.ethereum.rlp.RLPOutput;
+import org.apache.tuweni.bytes.Bytes;
 
-import java.util.Objects;
-
-import org.apache.tuweni.bytes.Bytes48;
-
-/** This class contains the data for a KZG proof for a KZG commitment. */
-public class KZGProof {
-  final Bytes48 data;
-
+/**
+ * Represents a KZG proof, which contains data used in the KZG scheme. This interface defines the
+ * contract for a KZG proof, which provides access to its data.
+ */
+public interface KZGProof {
   /**
-   * Constructor for a KZG proof.
+   * Get the data of the Blob.
    *
-   * @param data The data for the KZG proof.
+   * @return the data.
    */
-  public KZGProof(final Bytes48 data) {
-    this.data = data;
-  }
-
-  /**
-   * Reads a KZG proof from the RLP input.
-   *
-   * @param input The RLP input.
-   * @return The KZG proof.
-   */
-  public static KZGProof readFrom(final RLPInput input) {
-    final Bytes48 bytes = input.readBytes48();
-    return new KZGProof(bytes);
-  }
-
-  /**
-   * Writes the KZG proof to the RLP output.
-   *
-   * @param out The RLP output.
-   */
-  public void writeTo(final RLPOutput out) {
-    out.writeBytes(data);
-  }
-
-  /**
-   * Gets the data for the KZG proof.
-   *
-   * @return The data for the KZG proof.
-   */
-  public Bytes48 getData() {
-    return data;
-  }
-
-  @Override
-  public boolean equals(final Object o) {
-    if (this == o) return true;
-    if (o == null || getClass() != o.getClass()) return false;
-    KZGProof kzgProof = (KZGProof) o;
-    return Objects.equals(getData(), kzgProof.getData());
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(getData());
-  }
+  Bytes getData();
 }
